Conflicts:
	build/jam/MiscRules

+++ src/kits/tracker/GroupedMenu.cpp
@@ -24,10 +24,10 @@ TMenuItemGroup::TMenuItemGroup(const char* name)
 TMenuItemGroup::~TMenuItemGroup()
 {
 	free((char*)fName);
-	
+
 	if (fMenu == NULL) {
 		BMenuItem* item;
-		while ((item = RemoveItem(0L)) != NULL)
+		while ((item = RemoveItem((int32)0)) != NULL)
 			delete item;
 	}
 }
@@ -176,7 +176,7 @@ TGroupedMenu::TGroupedMenu(const char* name)
 TGroupedMenu::~TGroupedMenu()
 {
 	TMenuItemGroup* group;
-	while ((group = static_cast<TMenuItemGroup*>(fGroups.RemoveItem(0L)))
+	while ((group = static_cast<TMenuItemGroup*>(fGroups.RemoveItem((int32)0)))
 			!= NULL) {
 		delete group;
 	}
